throws exception on delete mapping if user not found

+++ fullstack-backend/src/main/java/com/springreact/fullstackbackend/controller/UserController.java
@@ -55,8 +55,12 @@ public class UserController {
 	}
 	
 	@DeleteMapping("/users/{id}")
-	void deleteUser(@PathVariable Long id) {
+	String deleteUser(@PathVariable Long id) {
+		if(!userRepository.existsById(id)) {
+			throw new UserNotFoundException(id);
+		}
 		 userRepository.deleteById(id);
+		 return "User with id "+id+" was successfully deleted.";
 	}
